DFR-1338
